Reorganize sort module
